+++ app/routers/workplace.py
@@ -1,17 +1,13 @@
-import pathlib
 from typing import List
 from uuid import UUID
 
 from beanie import DeleteRules, WriteRules
 from beanie.operators import In, RegEx
-from fastapi import APIRouter, Body, Depends, Path, UploadFile, status
-from fastapi.responses import FileResponse
+from fastapi import APIRouter, Body, Depends, Path, status
 
-from app.auth.oauth2 import admin, get_current_user, guest, member
-from app.core.download import downloader
-from app.core.exceptions import WorkplaceFileNotFoundException
+from app.auth.oauth2 import admin, get_current_user, guest
 from app.schemas.documents import Role, User, UserAssignedWorkplace, Workplace
-from app.schemas.models import FileModelOut, SuccessfulResponse, WorkplaceCreation
+from app.schemas.models import SuccessfulResponse, WorkplaceCreation
 
 router = APIRouter(tags=["Workplace"])
 
@@ -52,28 +48,6 @@ async def delete_workplace(workplace_id: UUID = Path(...), user: UserAssignedWor
     workplace = await Workplace.find_one(Workplace.id == workplace_id, fetch_links=True)
     await workplace.delete(link_rule=DeleteRules.DELETE_LINKS)
     return None
-
-
-@router.post("/workplaces/{workplace_id}/file", status_code=status.HTTP_201_CREATED, response_model=FileModelOut)
-async def add_file(
-    file_to_upload: UploadFile,
-    workplace_id: UUID = Path(...),
-    user: UserAssignedWorkplace = Depends(member),
-):
-    filename: str = await downloader(file_to_upload, workplace_id)
-    file_url = f"/workplaces/{workplace_id}/file/{filename}"
-    return FileModelOut(url=file_url)
-
-
-@router.get("/workplaces/{workplace_id}/file/{filename}", status_code=status.HTTP_200_OK, response_class=FileResponse)
-async def get_file(
-    workplace_id: UUID = Path(...), filename: str = Path(...), user: UserAssignedWorkplace = Depends(member)
-):
-    local_storage = pathlib.Path(__file__).parent.parent.parent.resolve()
-    path_file = local_storage.joinpath(pathlib.Path(f"assets/{workplace_id}/{filename}"))
-    if not pathlib.Path.is_file(path_file):
-        raise WorkplaceFileNotFoundException("Файл не найден")
-    return FileResponse(path_file)
 
 
 @router.get(
